factored button thread to button.c
