+++ src/main.c
@@ -17,21 +17,13 @@
 #include "hog.h"
 #include "revsvc.h"
 #include "batterylvl.h"
+#include "button.h"
 
 static int bond_count;
 
 #define LED0_NODE DT_ALIAS(led0)
-#define SW3_NODE DT_ALIAS(sw3)
-
 static const struct gpio_dt_spec led = GPIO_DT_SPEC_GET(LED0_NODE, gpios);
-static const struct gpio_dt_spec sw3 = GPIO_DT_SPEC_GET(SW3_NODE, gpios);
 
-#define BUTTON_THREAD_STACK_SIZE 1024
-#define BUTTON_THREAD_PRIORITY 5
-
-K_THREAD_STACK_DEFINE(button_thread_stack, BUTTON_THREAD_STACK_SIZE);
-struct k_thread button_thread_data;
-k_tid_t button_thread_tid;
 
 LOG_MODULE_REGISTER(main, LOG_LEVEL_DBG);
 
@@ -42,13 +34,13 @@ static void advertising_start(struct k_work *work);
 static K_WORK_DEFINE(start_advertising_worker, advertising_start);
 
 // Define thread stack sizes and priorities
-// #define REV_SVC_THREAD_STACK_SIZE 1024  // Adjust based on requirements
+
 #define HOG_BUTTON_THREAD_STACK_SIZE 1024  // Adjust based on requirements
-// #define REV_SVC_THREAD_PRIORITY 5  // Priority for rev_svc_loop thread
+
 #define HOG_BUTTON_THREAD_PRIORITY 5  // Priority for hog_button_loop thread
 
 // Thread stack and control blocks
-// K_THREAD_STACK_DEFINE(rev_svc_stack, REV_SVC_THREAD_STACK_SIZE); 
+
 K_THREAD_STACK_DEFINE(hog_button_stack, HOG_BUTTON_THREAD_STACK_SIZE); 
 // static struct k_thread rev_svc_thread_data;
 static struct k_thread hog_button_thread_data;
@@ -137,60 +129,10 @@ void advertise_with_acceptlist(struct k_work *work)
 K_WORK_DEFINE(advertise_acceptlist_work, advertise_with_acceptlist);
 
 
-void button_thread_fn(void *arg1, void *arg2, void *arg3) {
-    int ret;
-
-    if (!device_is_ready(sw3.port)) {
-        LOG_ERR("Button device %s not ready", sw3.port->name);
-        return;
-    }
-
-    ret = gpio_pin_configure_dt(&sw3, GPIO_INPUT);
-    if (ret < 0) {
-        LOG_ERR("Failed to configure button GPIO (err %d)", ret);
-        return;
-    }
-
-    // Enable pull-up or pull-down (if needed)
-    ret = gpio_pin_interrupt_configure_dt(&sw3, GPIO_INT_EDGE_TO_ACTIVE);
-    if (ret < 0) {
-        LOG_ERR("Failed to configure button interrupt (err %d)", ret);
-        return;
-    }
-
-    LOG_INF("Button thread started");
-
-    while (1) {
-        if (gpio_pin_get_dt(&sw3)) {
-            LOG_INF("Button pressed! Initiating Bluetooth unpair...");
-            bt_unpair(BT_ID_DEFAULT, BT_ADDR_LE_ANY);
-
-            // Debounce delay
-            k_sleep(K_MSEC(500));
-        }
-
-        k_sleep(K_MSEC(100));  // Polling delay
-    }
-}
 
 
-// Thread entry function for rev_svc_loop
-void rev_svc_thread(void *arg1, void *arg2, void *arg3) {
-    ARG_UNUSED(arg1);
-    ARG_UNUSED(arg2);
-    ARG_UNUSED(arg3);
 
-    rev_svc_loop();
-}
 
-// Thread entry function for hog_button_loop
-void hog_button_thread(void *arg1, void *arg2, void *arg3) {
-    ARG_UNUSED(arg1);
-    ARG_UNUSED(arg2);
-    ARG_UNUSED(arg3);
-
-    hog_button_loop();
-}
 
 static struct bt_le_adv_param *adv_param_normal = BT_LE_ADV_PARAM(
 	(BT_LE_ADV_OPT_CONN |
@@ -334,16 +276,7 @@ int main(void)
 
 
 	bt_conn_auth_info_cb_register(&bt_conn_auth_info);
-	    // Create a thread for the rev_svc_loop function
- 
-    // Create thread for rev_svc_loop
-    // k_thread_create(&rev_svc_thread_data, rev_svc_stack,
-    //                 K_THREAD_STACK_SIZEOF(rev_svc_stack),
-    //                 rev_svc_thread,
-    //                 NULL, NULL, NULL,
-    //                 REV_SVC_THREAD_PRIORITY,
-    //                 0,
-    //                 K_NO_WAIT);
+
 
     LOG_INF("rev_svc_loop thread started\n");
 
@@ -356,13 +289,14 @@ int main(void)
                     0,
                     K_NO_WAIT);
 
-	    button_thread_tid = k_thread_create(&button_thread_data, button_thread_stack,
-                                        K_THREAD_STACK_SIZEOF(button_thread_stack),
-                                        button_thread_fn, NULL, NULL, NULL,
-                                        BUTTON_THREAD_PRIORITY, 0, K_NO_WAIT);
+	    // button_thread_tid = k_thread_create(&button_thread_data, button_thread_stack,
+        //                                 K_THREAD_STACK_SIZEOF(button_thread_stack),
+        //                                 button_thread_fn, NULL, NULL, NULL,
+        //                                 BUTTON_THREAD_PRIORITY, 0, K_NO_WAIT);
 
 	batteryThreadinit();
 	rev_svc_thread_init();
+	rev_button_thread_init();
 	// hog_button_loop();
 
 	while (1) {
